refactor : myScheduleDetail.js, postScheduleDetail.js url alter 텍스트 수정
refactor : PostScheduleRestController,
PostScheduleService arg수정

feat : PostScheduleDetail.html
삭제하기 button 조건 추가

+++ src/main/java/com/travel/letsgospringboot/postschedule/service/PostScheduleService.java
@@ -115,9 +115,12 @@ public class PostScheduleService {
     }
 
     @Transactional
-    public void deletePostSchedule(String postId) {
+    public void deletePostSchedule(String postId, String loingUserId) {
         postScheduleRepository.deleteVisitItem(postId);
-        postScheduleRepository.deleteSchedulePost(postId);
+        String writerId = getUserId(postId);
+        if(writerId.equals(loingUserId)){
+            postScheduleRepository.deleteSchedulePost(postId);
+        }
     }
 
     @Transactional
